requirements.txt
numpy
PyOpenGL
glfw